refactor: address daily quality findings for 2026-08-07
Split MorphingTabs geometry/parts, modularize grant overlay panel and
position helpers, and extract Run log ANSI stripping into run_log_ansi.
